+++ web/src/test/kotlin/gruzilkin/iot/controllers/ChartsControllerTest.kt
@@ -14,6 +14,7 @@ import org.springframework.security.test.web.servlet.request.SecurityMockMvcRequ
 import org.springframework.test.context.jdbc.Sql
 import org.springframework.test.web.servlet.MockMvc
 import org.springframework.test.web.servlet.get
+import java.math.BigDecimal
 import java.security.Principal
 
 //@Disabled
@@ -37,8 +38,8 @@ class ChartsControllerTest {
     @WithMockUser(username = "1")
     fun index() {
         val device = deviceService.save(Principal { "1" }, "device_1")
-        sensorDataService.saveSensorData(SensorDataEvent(device.id!!, "temperature", 1.0, 1))
-        sensorDataService.saveSensorData(SensorDataEvent(device.id!!, "temperature", 2.0, 2))
+        sensorDataService.saveSensorData(SensorDataEvent(device.id!!, "temperature", BigDecimal.valueOf(1), 1))
+        sensorDataService.saveSensorData(SensorDataEvent(device.id!!, "temperature", BigDecimal.valueOf(2), 2))
 
         mockMvc.get("/charts/${device.id}") {
             with(csrf())
